Further migrations and simplificaitons
