+++ datafusion/physical-optimizer/tests/mod.rs
@@ -17,6 +17,4 @@
 
 mod enforce_sorting;
 mod limited_distinct_aggregation;
-mod replace_with_order_preserving_variants;
 mod sanity_checker;
-mod test_utils;
